# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FAR Mail API
  version: 1.0.0
  description: |
    FAR Mail 的公开接口与自动化 API。服务器地址使用相对路径，始终指向
    当前部署来源，不包含写死的生产域名或 IP。

    `/api/v1` 仅接受专用 API Token 的 `Authorization: Bearer`；
    Admin Key 仅供 `/console/v1` 使用，不能替代 API Token。
servers:
  - url: /
    description: 当前部署来源
...
paths:
...
  /api/v1/mailboxes:
    get:
      tags: [Mailboxes]
      operationId: listMailboxes
      summary: 查询当前 Token 可访问的邮箱
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: q
          in: query
          schema: { type: string }
          description: 邮箱地址搜索词
        - name: domain
          in: query
          schema: { type: string }
          description: 限定根域或子域
        - name: keep_forever
          in: query
          schema: { type: boolean }
        - name: expiring_within_hours
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: 分页邮箱列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailboxPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Mailboxes]
      operationId: createMailbox
      summary: 创建邮箱
      description: address 与 domain 均可省略；domain 可为活动根域或其有效子域。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMailbox"
      responses:
        "201":
          description: 邮箱已创建
          content:
            application/json:
              schema:
                type: object
                required: [mailbox]
                properties:
                  mailbox:
                    $ref: "#/components/schemas/Mailbox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/mailboxes/{id}:
    parameters:
      - $ref: "#/components/parameters/MailboxID"
    get:
      tags: [Mailboxes]
      operationId: getMailbox
      summary: 读取邮箱
      responses:
        "200":
          description: 邮箱详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  mailbox:
                    $ref: "#/components/schemas/Mailbox"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Mailboxes]
      operationId: deleteMailbox
      summary: 删除邮箱及关联邮件
      responses:
        "200":
          description: 删除完成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 32 lowercase hex characters
      description: 专用 API Token。不能使用 Admin Key，也不接受 X-API-Key 或 query-string key。
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    MailboxID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: 请求格式或参数无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 缺少凭据或凭据无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Token scope 不允许该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或当前 Token 不可访问
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        success: { type: boolean, const: false }
        error: { type: string }
        message: { type: string }
        error_code: { type: string }
        status: { type: integer }
        request_id: { type: string, format: uuid }
...
    CreateMailbox:
      type: object
      properties:
        address:
          type: string
          description: 邮箱本地部分；省略时自动生成
        domain:
          type: string
          description: 活动根域或其子域；省略时随机选择根域
    Mailbox:
      type: object
      required: [id, address, domain_id, full_address, keep_forever, created_at]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        address: { type: string }
        domain_id: { type: integer, format: int64 }
        full_address: { type: string, format: email }
        expires_at: { type: [string, "null"], format: date-time }
        keep_forever: { type: boolean }
        email_count: { type: integer }
        latest_received_at: { type: [string, "null"], format: date-time }
        latest_code: { type: string }
        latest_link: { type: string }
        created_at: { type: string, format: date-time }
    MailboxPage:
      type: object
      required: [data, total, page, size]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Mailbox"
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        q: { type: string }
        domain: { type: string }
        keep_forever: { type: boolean }
        expiring_within_hours: { type: integer }
...
    DeleteResult:
      type: object
      additionalProperties: true
      properties:
        message: { type: string }
        deleted: { type: integer }